#23 Preparation - Moved the EntraId token acquisition classes to EntraId namespace
