+++ src/ChatSuite.Sdk/ChatSuite.Sdk/Plugins/Security/EntraId/EntraIdDaemonAccessTokenProvider.cs
@@ -1,7 +1,6 @@
-﻿using ChatSuite.Sdk.Plugin.Security;
-using Microsoft.Extensions.Options;
+﻿using Microsoft.Extensions.Options;
 
-namespace ChatSuite.Sdk.Plugins.Security;
+namespace ChatSuite.Sdk.Plugins.Security.EntraId;
 
 internal class EntraIdDaemonAccessTokenProvider(
 	IOptions<EntraIdDaemonTokenAcquisitionSettings> entraIDDaemonTokenAcquisitionSettings,
